+++ src/main/java/ladysnake/effective/mixin/WaterFluidMixin.java
@@ -33,12 +33,31 @@ public class WaterFluidMixin {
         return false;
     }
 
+    @Unique
+    private static boolean shouldRipple(World world, BlockPos pos) {
+        if (EffectiveConfig.rainRippleDensity > 0) {
+            FluidState fluidState = world.getFluidState(pos);
+            if (fluidState.isStill() && world.isRaining() && world.getBlockState(pos.add(0, 1, 0)).isAir()) {
+                return true;
+            }
+        }
+        return false;
+    }
+
     @Inject(method = "randomDisplayTick", at = @At("HEAD"))
-    protected void effective$splash(World world, BlockPos pos, FluidState state, net.minecraft.util.math.random.Random random, CallbackInfo ci) {
+    protected void effective$splashAndRainRipples(World world, BlockPos pos, FluidState state, net.minecraft.util.math.random.Random random, CallbackInfo ci) {
+        // flowing water splashes
         if (shouldSplash(world, pos.up())) {
             Vec3d vec3d = state.getVelocity(world, pos);
             for (int i = 0; i < random.nextInt(50); i++) {
                 world.addParticle(ParticleTypes.SPLASH, pos.getX() + .5 + random.nextGaussian() / 2f, pos.getY() + 1 + random.nextFloat(), pos.getZ() + .5 + random.nextGaussian() / 2f, vec3d.getX() * random.nextFloat(), random.nextFloat() / 10f, vec3d.getZ() * random.nextFloat());
+            }
+        }
+
+        // still water rain ripples
+        if (shouldRipple(world, pos)) {
+            for (int i = 0; i <= random.nextInt(EffectiveConfig.rainRippleDensity); i++) {
+                world.addParticle(Effective.RIPPLE, pos.getX() + .5 + random.nextGaussian() / 2f, pos.getY() + 0.9f, pos.getZ() + .5 + random.nextGaussian() / 2f, 0f, 0f, 0f);
             }
         }
     }